pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "txtguy"
version = "0.1.0"
description = "A versatile text processing tool with LLM integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Billy Wang", email = "billy@techxartisan.com"}
]
keywords = ["text", "document", "file", "llm", "summarization", "ai", "transcript", "summary", "srt", "subtitle", "rename", "processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "click>=8.0.0",
    "charset-normalizer>=3.0.0",
    "pysrt>=1.1.2",
    "webvtt-py>=0.4.6",
]

[project.scripts]
txtguy = "txtguy.cli:main"

[project.urls]
Homepage = "https://github.com/youyoubilly/txt-sum"
Repository = "https://github.com/youyoubilly/txt-sum"
Issues = "https://github.com/youyoubilly/txt-sum/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["txtguy*"]

